add card and image
