+++ apps/api-journeys/src/modules/block/block.spec.ts
@@ -39,99 +39,6 @@ describe('BlockModule', () => {
     })
   }
 
-  it('returns CardBlock', async () => {
-    const parentBlockId = uuidv4()
-    const coverBlockId = uuidv4()
-    const card: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'CardBlock',
-      parentBlockId,
-      parentOrder: 0,
-      extraAttrs: {
-        backgroundColor: '#FFF',
-        coverBlockId,
-        themeMode: ThemeMode.light,
-        themeName: ThemeName.base
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([card])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on CardBlock {
-              backgroundColor
-              coverBlockId
-              themeMode
-              themeName
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: card.id,
-        __typename: 'CardBlock',
-        parentBlockId,
-        backgroundColor: '#FFF',
-        coverBlockId,
-        themeMode: ThemeMode.light,
-        themeName: ThemeName.base
-      }
-    ])
-  })
-
-  it('returns ImageBlock', async () => {
-    const parentBlockId = uuidv4()
-    const image: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'ImageBlock',
-      parentBlockId,
-      parentOrder: 2,
-      extraAttrs: {
-        src: 'https://source.unsplash.com/random/1920x1080',
-        alt: 'random image from unsplash',
-        width: 1920,
-        height: 1080
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([image])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on ImageBlock {
-              src
-              alt
-              width
-              height
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: image.id,
-        __typename: 'ImageBlock',
-        parentBlockId,
-        src: 'https://source.unsplash.com/random/1920x1080',
-        alt: 'random image from unsplash',
-        width: 1920,
-        height: 1080
-      }
-    ])
-  })
-
   it('returns StepBlock', async () => {
     const parentBlockId = uuidv4()
     const nextBlockId = uuidv4()
