+++ apps/api-journeys/src/modules/index.ts
@@ -1,7 +1,9 @@
 export { actionModule } from './action'
 export { blockModule } from './block'
 export { buttonModule } from './button'
+export { cardModule } from './card'
 export { iconModule } from './icon'
+export { imageModule } from './image'
 export { journeyModule } from './journey'
 export { radioQuestionModule } from './radioQuestion'
 export { responseModule } from './response'
